feat: add Emby support via shared media-server layer
Emby and Jellyfin share one API, so the Jellyfin integration is generalized
into a serviceId-parameterized media-server layer (services, hooks, tab screen,
now-playing widget) that both render through, rather than duplicated. The only
wire difference — image sizing params (fillWidth/fillHeight vs maxWidth/
maxHeight) — lives in lib/media-server-config.ts. Jellyfin behavior is unchanged
(serviceId defaults to jellyfin everywhere).

Adds the Emby tab, now-playing dashboard widget, service registration, auth
(X-Emby-Token), demo data, logo, and backend wiring. Closes #10.
