+++ components/dashboard/widget-registry.tsx
@@ -9,6 +9,7 @@ import {
   Gauge,
   HeartPulse,
   Inbox,
+  MonitorPlay,
   Newspaper,
   PlayCircle,
   Power,
@@ -27,6 +28,7 @@ import { TautulliActivityCard } from "@/components/dashboard/tautulli-activity-c
 import { OverseerrRequestsCard } from "@/components/dashboard/overseerr-requests-card";
 import { PlexNowPlayingCard } from "@/components/dashboard/plex-now-playing-card";
 import { JellyfinNowPlayingCard } from "@/components/dashboard/jellyfin-now-playing-card";
+import { EmbyNowPlayingCard } from "@/components/dashboard/emby-now-playing-card";
 import { ProwlarrStatsCard } from "@/components/dashboard/prowlarr-stats-card";
 import { BazarrWantedCard } from "@/components/dashboard/bazarr-wanted-card";
 import { WolDevicesCard } from "@/components/dashboard/wol-devices-card";
@@ -70,6 +72,11 @@ import {
   JELLYFIN_NOW_PLAYING_DEFAULT_SETTINGS,
   type JellyfinNowPlayingSettingsValue,
 } from "@/components/dashboard/widget-settings/jellyfin-now-playing-settings";
+import {
+  EmbyNowPlayingSettings,
+  EMBY_NOW_PLAYING_DEFAULT_SETTINGS,
+  type EmbyNowPlayingSettingsValue,
+} from "@/components/dashboard/widget-settings/emby-now-playing-settings";
 import {
   TautulliActivitySettings,
   TAUTULLI_ACTIVITY_DEFAULT_SETTINGS,
@@ -291,6 +298,16 @@ export const WIDGET_REGISTRY: Record<WidgetId, WidgetDefinition> = {
     settingsComponent: JellyfinNowPlayingSettings,
     defaultSettings: JELLYFIN_NOW_PLAYING_DEFAULT_SETTINGS,
   },
+  "emby-now-playing": {
+    id: "emby-now-playing",
+    label: "Emby Now Playing",
+    description: "Live playback sessions on your Emby server",
+    icon: MonitorPlay,
+    service: "emby",
+    component: EmbyNowPlayingCard,
+    settingsComponent: EmbyNowPlayingSettings,
+    defaultSettings: EMBY_NOW_PLAYING_DEFAULT_SETTINGS,
+  },
   "prowlarr-stats": {
     id: "prowlarr-stats",
     label: "Prowlarr Stats",
@@ -338,6 +355,7 @@ export type {
   NzbgetQueueSettingsValue,
   PlexNowPlayingSettingsValue,
   JellyfinNowPlayingSettingsValue,
+  EmbyNowPlayingSettingsValue,
   TautulliActivitySettingsValue,
   OverseerrRequestsSettingsValue,
   SpeedStatsSettingsValue,
