+++ docs/index.html
@@ -287,6 +287,10 @@
         <p>Now playing, recently added, continue watching</p>
       </div>
       <div class="service-card">
+        <h3>Emby</h3>
+        <p>Now playing, recently added, continue watching</p>
+      </div>
+      <div class="service-card">
         <h3>Glances</h3>
         <p>Server CPU, RAM, disk, network stats</p>
       </div>
