+++ lib/constants.ts
@@ -9,6 +9,7 @@ export const SERVICE_IDS = [
   "prowlarr",
   "plex",
   "jellyfin",
+  "emby",
   "glances",
   "bazarr",
 ] as const;
@@ -40,6 +41,9 @@ export const SERVICE_DEFAULTS: Record<
   prowlarr: { name: "Prowlarr", defaultPort: 9696, apiBasePath: "/api/v1", pingPath: "/system/status" },
   plex: { name: "Plex", defaultPort: 32400, apiBasePath: "", pingPath: "/identity" },
   jellyfin: { name: "Jellyfin", defaultPort: 8096, apiBasePath: "", pingPath: "/System/Info/Public" },
+  // Emby shares Jellyfin's API surface (same default port, root API path, and
+  // public System/Info endpoint). See lib/media-server-config.ts.
+  emby: { name: "Emby", defaultPort: 8096, apiBasePath: "", pingPath: "/System/Info/Public" },
   glances: { name: "Glances", defaultPort: 61208, apiBasePath: "/api/4", pingPath: "/cpu" },
   bazarr: { name: "Bazarr", defaultPort: 6767, apiBasePath: "/api", pingPath: "/system/status" },
 };
@@ -65,6 +69,7 @@ export const DASHBOARD_WIDGET_IDS = [
   "overseerr-requests",
   "plex-now-playing",
   "jellyfin-now-playing",
+  "emby-now-playing",
   "prowlarr-stats",
   "bazarr-wanted",
   "wol-devices",
